combined front and backend into one app
